Datagovmy/feat/dc revamp with index (#407)
* initial timeseries revamp

* wip: chart and table preview

* adjust styling of the preview card

* fix: fix styling of the preview card

* scroll to top of chart when change chart

* change embed to use new api and implementation

* fix: fix missing required prop for embed

* change DC index page list to card style

* fix: fix some styling for the DC index card

* fix: change truncate to display tooltip

* change to usestate

* add sidebar to DC show page

* fix: initial selected value for sidebar and update selected when toggle lang

* add related datasets in DC show

* add alternate styling to dc catalogue card

* remove unused ref

* add related dataset, and fix sidebar styling

* add sitewide banner in dgmy

* fix: fix i18n key

* change table preview to icon and hide related datasets if none

* add query params for visual state

* change banner to use full-width

* only use visual query as initial load

* refactor: move dc components to datagovmy-ui and apply accordingly in dmgy and opendosm

* tweak styling for dc index

* use component from dc for data-request

* move opendosm-pub gui to opendosm
