+++ apps/app/pages/gui/index.tsx
@@ -19,7 +19,7 @@ const GUIPage: Page = ({ meta }: InferGetStaticPropsType<typeof getStaticProps>)
 
 export const getStaticProps: GetStaticProps = withi18n("gui-opendosm-pub", async () => {
   return {
-    notFound: false,
+    notFound: true,
     props: {
       meta: {
         id: "gui",
