+++ apps/app/tsconfig.json
@@ -8,7 +8,6 @@
       "@lib/*": ["lib/*"],
       "@hooks/*": ["hooks/*"],
       "@dashboards/*": ["dashboards/*"],
-      "@data-catalogue/*": ["data-catalogue/*"],
       "@misc/*": ["misc/*"],
       "@config/*": ["config/*"]
     }
