+++ apps/opendosm/lib/routes.ts
@@ -24,6 +24,9 @@ export const routes = {
   WHOLESALE_RETAIL: "/dashboard/wholesale-retail-trade",
   WELLBEING: "/dashboard/wellbeing",
   PUBLICATIONS: "/publications",
+  GUI: "/gui",
+  GUI_OPENDOSM_PUB: "/gui/opendosm-pub",
+  GUI_CATALOGUE: "/gui/catalogue",
 };
 
 export const static_routes: string[] = (() => {
