+++ apps/opendosm/pages/gui/index.tsx
@@ -3,15 +3,15 @@ import { Page } from "datagovmy-ui/types";
 import { Metadata } from "datagovmy-ui/components";
 import { useTranslation } from "datagovmy-ui/hooks";
 import { withi18n } from "datagovmy-ui/decorators";
-import GUILayout from "@misc/gui/layout";
+import GUILayout from "misc/gui/layout";
 
-const GUICatalogue: Page = ({ meta }: InferGetStaticPropsType<typeof getStaticProps>) => {
+const GUIPage: Page = ({ meta }: InferGetStaticPropsType<typeof getStaticProps>) => {
   const { t } = useTranslation(["gui-opendosm-pub"]);
   return (
     <>
       <Metadata title={t("header")} description={t("description")} keywords={""} />
       <GUILayout>
-        <h1>hello from catalogue</h1>
+        <></>
       </GUILayout>
     </>
   );
@@ -19,7 +19,7 @@ const GUICatalogue: Page = ({ meta }: InferGetStaticPropsType<typeof getStaticPr
 
 export const getStaticProps: GetStaticProps = withi18n("gui-opendosm-pub", async () => {
   return {
-    notFound: true,
+    notFound: false,
     props: {
       meta: {
         id: "gui",
@@ -31,4 +31,4 @@ export const getStaticProps: GetStaticProps = withi18n("gui-opendosm-pub", async
   };
 });
 
-export default GUICatalogue;
+export default GUIPage;
