+++ packages/datagovmy-ui/src/components/Sidebar/index.tsx
@@ -3,17 +3,29 @@ import { Transition } from "@headlessui/react";
 import Button from "../Button";
 import { Bars3BottomLeftIcon, XMarkIcon } from "@heroicons/react/24/outline";
 import { useTranslation } from "../../hooks/useTranslation";
+import { clx } from "../../lib/helpers";
 
 interface SidebarProps {
   children: ReactNode;
   categories: Array<[category: string, subcategory: string[]]>;
   onSelect: (index: string) => void;
+  sidebarTitle?: string;
+  mobileClassName?: string;
+  initialSelected?: string;
 }
 
-const Sidebar: FunctionComponent<SidebarProps> = ({ children, categories, onSelect }) => {
+const Sidebar: FunctionComponent<SidebarProps> = ({
+  children,
+  categories,
+  onSelect,
+  sidebarTitle,
+  mobileClassName,
+  initialSelected,
+}) => {
   const { t } = useTranslation(["catalogue", "common"]);
-  const [selected, setSelected] = useState<string>();
-  const [mainFlag, setMainFlag] = useState<boolean>(false);
+  const [selected, setSelected] = useState<string>(
+    initialSelected ?? Array.isArray(categories[0]) ? categories[0][0] : categories[0]
+  );
   const [show, setShow] = useState<boolean>(false);
   const styles = {
     base: "px-4 lg:px-5 py-1.5 w-full rounded-none text-start leading-tight",
@@ -26,43 +38,39 @@ const Sidebar: FunctionComponent<SidebarProps> = ({ children, categories, onSele
     <>
       <div className="flex w-full flex-row">
         {/* Desktop */}
-        <div className="dark:border-r-washed-dark hidden border-r pb-6 lg:block lg:w-1/4 xl:w-1/5">
+        <div className="dark:border-r-washed-dark hidden border-r pb-6 lg:block lg:w-1/5 xl:w-1/5">
           <ul className="sticky top-14 flex h-[calc(100vh-56px)] flex-col gap-2 overflow-auto pt-3">
             <li>
-              <h5 className={styles.base}>{t("category")}</h5>
+              <h5 className={styles.base}>{sidebarTitle ?? t("category")}</h5>
             </li>
             {categories.length > 0 ? (
-              categories.map(([category, subcategory]) => (
+              categories.map(([category, subcategory], index) => (
                 <li key={`${category}: ${subcategory[0]}`} title={category}>
                   <Button
                     className={[
                       styles.base,
-                      mainFlag && selected === `${category}: ${subcategory[0]}`
-                        ? styles.active
-                        : styles.default,
+                      selected === `${category}` ? styles.active : styles.default,
                     ].join(" ")}
                     onClick={() => {
-                      setSelected(`${category}: ${subcategory[0]}`);
-                      setMainFlag(true);
-                      onSelect(`${category}: ${subcategory[0]}`);
+                      setSelected(`${category}`);
+                      onSelect(
+                        subcategory.length > 0 ? `${category}: ${subcategory[0]}` : `${category}`
+                      );
                     }}
                   >
                     {category}
                   </Button>
                   <ul className="ml-5 space-y-1">
-                    {subcategory.length &&
-                      subcategory.map(title => (
+                    {Boolean(subcategory.length) &&
+                      subcategory.map((title, subIndex) => (
                         <li key={title} title={title}>
                           <Button
                             className={[
                               styles.base,
-                              !mainFlag && selected === `${category}: ${title}`
-                                ? styles.active
-                                : styles.default,
+                              selected === `${category}: ${title}` ? styles.active : styles.default,
                             ].join(" ")}
                             onClick={() => {
                               setSelected(`${category}: ${title}`);
-                              setMainFlag(false);
                               onSelect(`${category}: ${title}`);
                             }}
                           >
@@ -82,15 +90,20 @@ const Sidebar: FunctionComponent<SidebarProps> = ({ children, categories, onSele
         </div>
 
         {/* Mobile */}
-        <div className="relative w-full">
+        <div className="relative w-full lg:w-4/5 xl:w-4/5">
           <>
-            <div className="absolute top-[72px] block lg:hidden">
+            <div
+              className={clx(
+                "absolute block lg:hidden",
+                mobileClassName ? mobileClassName : "top-[72px]"
+              )}
+            >
               <Button
                 className="btn-default shadow-button"
                 icon={<Bars3BottomLeftIcon className="h-4 w-4" />}
                 onClick={() => setShow(true)}
               >
-                {t("category")}
+                {sidebarTitle ?? t("category")}
               </Button>
             </div>
             <Transition
@@ -106,7 +119,7 @@ const Sidebar: FunctionComponent<SidebarProps> = ({ children, categories, onSele
             >
               <ul className="flex flex-col gap-1 overflow-auto pt-2">
                 <li className="flex items-baseline justify-between">
-                  <h5 className={styles.base}>{t("category")}</h5>
+                  <h5 className={styles.base}>{sidebarTitle ?? t("category")}</h5>
 
                   <Button
                     className="hover:bg-washed dark:hover:bg-washed-dark group absolute right-2 top-2 flex h-8 w-8 items-center rounded-full"
@@ -126,13 +139,17 @@ const Sidebar: FunctionComponent<SidebarProps> = ({ children, categories, onSele
                         ].join(" ")}
                         onClick={() => {
                           setSelected(category);
-                          onSelect(`${category}: ${subcategory[0]}`);
+                          onSelect(
+                            subcategory.length > 0
+                              ? `${category}: ${subcategory[0]}`
+                              : `${category}`
+                          );
                         }}
                       >
                         {category}
                       </Button>
                       <ul className="ml-4">
-                        {subcategory.length &&
+                        {Boolean(subcategory.length) &&
                           subcategory.map(title => (
                             <li key={title}>
                               <Button
