+++ packages/datagovmy-ui/src/data-catalogue/index.ts
@@ -1,0 +1,12 @@
+export { default as CatalogueIndex } from "./Main";
+export { default as CatalogueCard } from "./Card";
+export { default as CataloguePreview } from "./Preview";
+export { default as CatalogueShow } from "./Show";
+export { default as CatalogueWidget } from "./Widget";
+
+// Type export
+export type { Catalogue } from "./Main";
+export type {} from "./Card";
+export type {} from "./Preview";
+export type { IDataViz } from "./Show";
+export type {} from "./Widget";
