+++ packages/datagovmy-ui/types/data-catalogue.d.ts
@@ -1,0 +1,1 @@
+export * from "datagovmy-ui/src/data-catalogue";
